custom-elements/ElementInternals-role.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body></body>

<script>
function conditionally_register_element(role) {
    if (!customElements.get(`${role}-element`)) {
        customElements.define(`${role}-element`, class extends HTMLElement {
            constructor() {
                super();
                this.internals_ = this.attachInternals();
                this.internals_.role = role;
                this.internals_.ariaLabel = "Test label";
            }
        });
    }
}

function generate_role_element(role) {
    conditionally_register_element(role);
    const el = document.createElement(`${role}-element`);
    document.body.append(el);
    return el;
}

function generate_child_role_element(role, parentElement) {
    conditionally_register_element(role);
    const el = document.createElement(`${role}-element`);
    parentElement.append(el);
    return el;
}

const test_roles = [
    'alert',
    'alertdialog',
    'application',
    'article',
    'banner',
    'button',
    'cell',
    'checkbox',
    'combobox',
    'complementary',
    'contentinfo',
    'definition',
    'dialog',
    ['directory', 'list'], // is this expected?
    'document',
    'figure',
    'form',
    'grid',
    'group',
    'heading',
    ['img', 'image'],
    'feed',
    'link',
    'list',
    'listbox',
    'log',
    'main',
    'marquee',
    'math',
    'menu',
    'menubar',
    'meter',
    'navigation',
    'note',
    'progressbar',
    'radio',
    'radiogroup',
    'region',
    'scrollbar',
    'search',
    'searchbox',
    'separator',
    'slider',
    'spinbutton',
    'status',
    'switch',
    'table',
    'tablist',
    'tabpanel',
    'term',
    'textbox',
    'timer',
    'toolbar',
    'tooltip',
    'tree',
    'treegrid'
];

test_roles.map((testdata) => {
    let role;
    let roleName;
    if (Array.isArray(testdata)) {
        role = testdata[0];
        roleName = testdata[1];
    } else {
        role = testdata;
        roleName = testdata;
    }
    promise_test(async () => {
        const el = generate_role_element(role);
        const computedRole = await test_driver.get_computed_role(el);
        assert_equals(computedRole, roleName, el);
        el.remove();
    }, `Applies "${role}" via Element Internals`);
});

const child_test_roles = [
    ['table', 'row', 'columnheader'],
    ['grid', 'row', 'gridcell'],
    ['list', 'listitem'],
    ['menu', 'menuitem'],
    ['menu', 'menuitemcheckbox'],
    ['menu', 'menuitemradio'],
    ['listbox', 'option'],
    ['table', 'row'],
    ['table', 'rowgroup'],
    ['table', 'row', 'rowheader'],
    ['tablist', 'tab'],
    ['tree', 'treeitem'],
];

child_test_roles.map((roles) => {
    promise_test(async () => {
        const els = roles.reduce((acc, role) => {
            const previous = acc.at(-1);
            if (previous) {
                acc.push(generate_child_role_element(role, previous))
            } else {
                acc.push(generate_role_element(role));
            }
            return acc;
        }, []);
        const computedRoles = [];
        for(let i=0; i < els.length; i++) {
            computedRoles.push(await test_driver.get_computed_role(els[i]));
        }
        computedRoles.map((role, i) => {
            assert_equals(role, roles[i], els[i]);
        })
        els[0].remove();
    }, `Applies parent/child relationship of "${roles.join('"/"')}" via Element Internals`);
});
</script>
